Updated App.js with fetch functionality
Can receive data from backend now. Only triggered when a marker is clicked on because fetch code is currently nested inside the onChildClick function. Feel free to move this around to another function if you want this to be triggered elsewhere
Also I've marked "replace this" for where you will want to take this data and convert it to the interface